pyproject.toml
[project]
name = "cv-extract-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-google>=0.1.1",
    "langchain-google-genai>=2.1.9",
    "langgraph>=0.6.3",
    "paddleocr>=3.1.0",
    "paddlepaddle>=3.1.0",
    "pillow>=11.3.0",
    "pydantic[email]>=2.11.7",
    "pymupdf>=1.26.3",
    "python-doctr>=1.0.0",
    "python-dotenv>=1.1.1",
    "setuptools>=80.9.0",
]
